Did kind of a looot, oops
Switched out vim-polyglot for treesitter, which still seems to be a WIP
for typescript, so kept some of those packages to make that more
manageable.
With treesitter, more syntax highlighting seems to be good for both ruby
and tsx, so dracula is in and palenight is out.

Removed some packages I just can't be botehred to use ):